% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyUtil.R
\name{augmentCancerDataSet}
\alias{augmentCancerDataSet}
\title{This function augments the cancer specific dataset interventions per patient}
\usage{
augmentCancerDataSet(
  cancerCohortDataTable,
  interventionsVector,
  drugVector,
  timeWindowForInterventions
)
}
\description{
This function augments the cancer specific dataset interventions per patient
}
