% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastCancerSpecificAnalysis.R
\name{breastCancerSpecificAnalysis}
\alias{breastCancerSpecificAnalysis}
\title{Execute breast cancer specific characterization analysis}
\usage{
breastCancerSpecificAnalysis(
  cohortName,
  cancerCohortDataTable,
  cancerSpecificVectors,
  databaseId,
  outputFolder,
  minCellCount
)
}
\description{
Execute breast cancer specific characterization analysis
}
