\name{CancerTreatmentCharacterization-package}
\alias{CancerTreatmentCharacterization-package}
\alias{CancerTreatmentCharacterization}
\docType{package}
\title{
\packageTitle{CancerTreatmentCharacterization}
}
\description{
\packageDescription{CancerTreatmentCharacterization}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CancerTreatmentCharacterization}
\packageIndices{CancerTreatmentCharacterization}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{CancerTreatmentCharacterization}

Maintainer: \packageMaintainer{CancerTreatmentCharacterization}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
