% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyUtil.R
\name{getBuildSqlFileName}
\alias{getBuildSqlFileName}
\title{This file is a R script contains utility functions for the study
This function returns the sql script file name that builds the cancer specific dataset based on the cohortId}
\usage{
getBuildSqlFileName(cohortId)
}
\description{
This file is a R script contains utility functions for the study
This function returns the sql script file name that builds the cancer specific dataset based on the cohortId
}
