% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyUtil.R
\name{getCancerDataSet}
\alias{getCancerDataSet}
\title{This function returns the appropriate data set needed to run the analysis for a specific cancer cohort based on the cohortId}
\usage{
getCancerDataSet(cohortDatabaseSchema, cohortId, connection)
}
\description{
This function returns the appropriate data set needed to run the analysis for a specific cancer cohort based on the cohortId
}
