% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyUtil.R
\name{getCancerDataSetName}
\alias{getCancerDataSetName}
\title{This function returns the cancer specific dataset name based on the cohortId}
\usage{
getCancerDataSetName(cohortId)
}
\description{
This function returns the cancer specific dataset name based on the cohortId
}
