% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyUtil.R
\name{getConceptListSqlFileName}
\alias{getConceptListSqlFileName}
\title{This function returns the sql script file name that creates a list of concepts for the specific cancer based on the cohortId}
\usage{
getConceptListSqlFileName(cohortId)
}
\description{
This function returns the sql script file name that creates a list of concepts for the specific cancer based on the cohortId
}
