% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyUtil.R
\name{getVectorsForSpecificCancer}
\alias{getVectorsForSpecificCancer}
\title{This function returns the appropriate vectors needed to run the analysis for a specific cancer cohort based on the cohortId}
\usage{
getVectorsForSpecificCancer(cohortId)
}
\description{
This function returns the appropriate vectors needed to run the analysis for a specific cancer cohort based on the cohortId
}
