% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProstateCancerSpecificAnalysis.R
\name{prostateCancerSpecificAnalysis}
\alias{prostateCancerSpecificAnalysis}
\title{Execute breast cancer specific characterization analysis}
\usage{
prostateCancerSpecificAnalysis(
  cohortName,
  cancerCohortDataTable,
  cancerSpecificVectors,
  outputFolder,
  minCellCount
)
}
\description{
Execute breast cancer specific characterization analysis
}
