% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortTreatmentAnalysis.R
\name{runCancerTreatmentAnalysis}
\alias{runCancerTreatmentAnalysis}
\title{Execute the cancer characterization analysis}
\usage{
runCancerTreatmentAnalysis(
  connection,
  cohortDatabaseSchema,
  cohortId,
  outputFolder,
  minCellCount = 5
)
}
\arguments{
\item{connection}{An object of type \code{connection} that has an established database connection.}

\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{minCellCount}{The minimum number of subjects contributing to a count before it can be included
in packaged results.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortTable}{The name of the table that will be created in the work database schema.
This table will hold the exposure and outcome cohorts used in this
study.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}

\item{databaseId}{A short string for identifying the database (e.g.
'Synpuf').}

\item{databaseName}{The full name of the database (e.g. 'Medicare Claims
Synthetic Public Use Files (SynPUFs)').}

\item{databaseDescription}{A short description (several sentences) of the database.}

\item{createCohorts}{Create the cohortTable table with the exposure and outcome cohorts?}
}
\description{
Execute the cancer characterization analysis
}
\details{
This function executes the cohort diagnostics.
}
